backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy>=2.0.35
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic[email]==2.6.4
python-dotenv==1.0.1
joblib>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
